bug fixes for main page, added bidding, started login page

+++ serverProject/server/Item.java
@@ -13,7 +13,7 @@ public class Item {
         this.time = time;
     }
 
-    public boolean bid(int bid){
+    public boolean bid(double bid){
         if(bid > price){
             price = bid;
             return true;
